refactor(engine): extract primer-engine lib crate for shared wiring helpers
Move build_backend / build_classifier / build_extractor / build_comprehension
/ build_fastembed_embedder / build_ollama_embedder, plus the slug + path
resolution + learner reconciliation helpers, from primer-cli/src/main.rs to
a new internal lib crate. Both binaries (primer-cli today, primer-gui next)
import from primer-engine — duplicating ~500 lines of wiring across two
binaries would invite divergence as backends/features evolve.

Pure code move: no behaviour change to the CLI. All 14 build_* dispatch
tests + 18 path/learner tests move with their helpers. Workspace tests
green; embedding + ollama-embedding feature combos build clean; stub
end-to-end smoke (boot → greeting → quit) succeeds.
