+++ src/crates/primer-cli/src/main.rs
@@ -17,34 +17,24 @@ use std::io::{self, BufRead, Write};
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 
-use chrono::Utc;
 use clap::Parser;
-use primer_classifier::{
-    ClassifierSettings, EngagementClassifier, LlmEngagementClassifier, StubEngagementClassifier,
-};
+use primer_classifier::{ClassifierSettings, EngagementClassifier};
 use primer_core::config::PedagogyConfig;
-use primer_core::error::{PrimerError, Result};
+use primer_core::error::PrimerError;
 use primer_core::i18n::{Locale, render_inference_error};
 use primer_core::inference::InferenceBackend;
 use primer_core::knowledge::KnowledgeBase;
-use primer_core::learner::*;
 use primer_core::storage::{LearnerStore, SessionStore};
-use primer_extractor::{
-    ConceptExtractor, ExtractorSettings, LlmConceptExtractor, StubConceptExtractor,
+use primer_engine::{
+    BackendParams, IN_MEMORY, build_backend, build_classifier, build_comprehension,
+    build_extractor, build_fastembed_embedder, build_ollama_embedder, create_learner_with_id,
+    reconcile_persisted_learner, resolve_session_db_path, should_show_first_run_banner,
+    verify_resume_locale_match,
 };
-use primer_inference::stub::StubBackend;
+use primer_extractor::{ConceptExtractor, ExtractorSettings};
 use primer_knowledge::SqliteKnowledgeBase;
 use primer_pedagogy::DialogueManager;
-use unicode_normalization::UnicodeNormalization;
 use uuid::Uuid;
-
-/// SQLite path token for an in-memory database — used as the default
-/// for `--knowledge-db` when no path is given. Sessions persist to a
-/// per-learner file under `~/.primer/` instead.
-const IN_MEMORY: &str = ":memory:";
-
-/// Subdirectory under `$HOME` for per-learner session databases.
-const PRIMER_HOME_DIR: &str = ".primer";
 
 #[derive(Parser, Debug)]
 #[command(name = "primer", about = "The Primer — a Socratic learning companion")]
@@ -267,473 +257,6 @@ fn parse_mic_silence_ms(s: &str) -> std::result::Result<u32, String> {
     Ok(n)
 }
 
-/// Slugify a learner name into a filesystem-safe filename stem.
-///
-/// The input is first NFC-normalized so two visually identical names
-/// (e.g. precomposed `é` vs decomposed `e` + combining acute) map to
-/// the same slug. Characters that Unicode classifies as alphanumeric
-/// — Latin, Cyrillic, CJK, etc. — are kept (Latin is lowercased; CJK
-/// has no case so it round-trips). Every other character becomes `-`;
-/// runs of `-` collapse; leading/trailing `-` are stripped. An empty
-/// result falls back to `default` so we always produce a valid filename.
-fn slug(name: &str) -> String {
-    let normalized: String = name.nfc().collect();
-    let lowered = normalized.to_lowercase();
-    let mut out = String::with_capacity(lowered.len());
-    let mut last_was_sep = true; // suppress leading sep
-    for c in lowered.chars() {
-        if c.is_alphanumeric() {
-            out.push(c);
-            last_was_sep = false;
-        } else if !last_was_sep {
-            out.push('-');
-            last_was_sep = true;
-        }
-    }
-    while out.ends_with('-') {
-        out.pop();
-    }
-    if out.is_empty() {
-        "default".to_string()
-    } else {
-        out
-    }
-}
-
-/// Resolve the path to use for the session database.
-/// `:memory:` when `no_persist` is set; otherwise the explicit path if
-/// given, falling back to `<home>/.primer/<slug(name)>.db`. The home
-/// directory is taken as a parameter so callers can supply it from any
-/// source (env var in production, synthetic value in tests) without
-/// this function touching the process environment.
-fn resolve_session_db_path(
-    explicit: Option<PathBuf>,
-    home: &Path,
-    learner_name: &str,
-    no_persist: bool,
-) -> PathBuf {
-    if no_persist {
-        return PathBuf::from(IN_MEMORY);
-    }
-    explicit.unwrap_or_else(|| {
-        home.join(PRIMER_HOME_DIR)
-            .join(format!("{}.db", slug(learner_name)))
-    })
-}
-
-/// Should we print the "we just started persisting your sessions"
-/// banner? True only when the session DB is at the default path AND
-/// the file did not exist before this run AND the user did not opt
-/// out via `--no-persist`. The banner answers the legitimate "where
-/// did my conversation go?" question that the silent default-path
-/// change would otherwise raise.
-fn should_show_first_run_banner(
-    explicit_session_db: bool,
-    no_persist: bool,
-    file_existed_before: bool,
-) -> bool {
-    !explicit_session_db && !no_persist && !file_existed_before
-}
-
-/// Parameters needed by `build_backend` and `build_classifier` that would
-/// otherwise require borrowing from `Cli`. Extracted early (before any partial
-/// moves of `Cli` fields) so the helpers can be called after those moves.
-struct BackendParams {
-    api_key: Option<String>,
-    ollama_url: String,
-    classifier_backend: Option<String>,
-    classifier_model: Option<String>,
-    extractor_backend: Option<String>,
-    extractor_model: Option<String>,
-    comprehension_backend: Option<String>,
-    comprehension_model: Option<String>,
-}
-
-/// Construct an `InferenceBackend` of the named type with the given model.
-///
-/// All three backend variants are synchronous at construction time; the
-/// function signature is `async` only for uniformity with `build_classifier`.
-async fn build_backend(
-    backend_name: &str,
-    model: String,
-    params: &BackendParams,
-) -> Result<Arc<dyn InferenceBackend>> {
-    match backend_name {
-        "stub" => Ok(Arc::new(StubBackend)),
-        "cloud" => {
-            let api_key = params.api_key.clone().ok_or_else(|| {
-                PrimerError::Inference(
-                    "--api-key or ANTHROPIC_API_KEY required for cloud backend".into(),
-                )
-            })?;
-            Ok(Arc::new(primer_inference::cloud::CloudBackend::new(
-                "https://api.anthropic.com".to_string(),
-                api_key,
-                model,
-            )))
-        }
-        "ollama" => Ok(Arc::new(primer_inference::ollama::OllamaBackend::new(
-            params.ollama_url.clone(),
-            model,
-        ))),
-        other => Err(PrimerError::Inference(
-            format!("unknown backend: {other}").into(),
-        )),
-    }
-}
-
-/// Construct the engagement classifier according to the dispatch matrix:
-///
-/// | main backend | --classifier-backend | --classifier-model | outcome                               |
-/// |--------------|----------------------|--------------------|---------------------------------------|
-/// | stub         | (unset)              | (any)              | StubEngagementClassifier              |
-/// | *            | "stub"               | (any)              | StubEngagementClassifier              |
-/// | *            | some(X)              | None               | error (model required for non-stub)   |
-/// | *            | some(X)              | some(M)            | LlmEngagementClassifier(new backend X, model M) |
-/// | non-stub     | (unset)              | None               | LlmEngagementClassifier(Arc::clone main) |
-/// | non-stub     | (unset)              | some(M)            | LlmEngagementClassifier(new backend same type, model M) |
-async fn build_classifier(
-    main_backend: Arc<dyn InferenceBackend>,
-    main_backend_name: &str,
-    main_model: &str,
-    params: &BackendParams,
-    settings: ClassifierSettings,
-) -> Result<Arc<dyn EngagementClassifier>> {
-    // Warn on a flag combination the user probably didn't mean: passing
-    // --classifier-model alongside an explicit --classifier-backend stub
-    // looks like the model name will take effect, but stub ignores it.
-    if matches!(params.classifier_backend.as_deref(), Some("stub"))
-        && params.classifier_model.is_some()
-    {
-        tracing::warn!(
-            model = ?params.classifier_model,
-            "--classifier-model ignored: --classifier-backend is stub (deterministic, no model)"
-        );
-    }
-
-    match (main_backend_name, params.classifier_backend.as_deref()) {
-        // Explicit stub override always wins regardless of main backend.
-        (_, Some("stub")) => Ok(Arc::new(StubEngagementClassifier::new())),
-
-        // Main backend is stub and no classifier backend specified → default to stub.
-        // (LlmEngagementClassifier wrapping a stub backend would silently return
-        // Unknown on every classify call — that's worse than a deterministic stub.)
-        ("stub", None) => Ok(Arc::new(StubEngagementClassifier::new())),
-
-        // Explicit non-stub classifier backend → need a model too.
-        (_, Some(cls_backend_name)) => {
-            let model = params.classifier_model.clone().ok_or_else(|| {
-                PrimerError::Inference(
-                    "--classifier-model is required when --classifier-backend is set to a non-stub backend".into(),
-                )
-            })?;
-            let cls_backend = build_backend(cls_backend_name, model.clone(), params).await?;
-            Ok(Arc::new(LlmEngagementClassifier::new(
-                cls_backend,
-                model,
-                settings,
-            )))
-        }
-
-        // No --classifier-backend specified, main is not stub.
-        (_, None) => {
-            match params.classifier_model.as_deref() {
-                // No model override → reuse the main backend via Arc::clone.
-                // The classifier identifier becomes "llm:<main-model>".
-                None => Ok(Arc::new(LlmEngagementClassifier::new(
-                    Arc::clone(&main_backend),
-                    main_model.to_string(),
-                    settings,
-                ))),
-                // Model override only → construct a fresh backend of the same TYPE
-                // with the override model, because InferenceBackend::generate() does
-                // not accept a model argument — model is baked in at construction.
-                Some(override_model) => {
-                    let cls_backend =
-                        build_backend(main_backend_name, override_model.to_string(), params)
-                            .await?;
-                    Ok(Arc::new(LlmEngagementClassifier::new(
-                        cls_backend,
-                        override_model.to_string(),
-                        settings,
-                    )))
-                }
-            }
-        }
-    }
-}
-
-/// Construct the concept extractor according to the same dispatch matrix
-/// as `build_classifier`:
-///
-/// | main backend | --extractor-backend | --extractor-model | outcome                                |
-/// |--------------|---------------------|-------------------|----------------------------------------|
-/// | stub         | (unset)             | (any)             | StubConceptExtractor                   |
-/// | *            | "stub"              | (any)             | StubConceptExtractor                   |
-/// | *            | some(X)             | None              | error (model required for non-stub)    |
-/// | *            | some(X)             | some(M)           | LlmConceptExtractor(new backend X, model M) |
-/// | non-stub     | (unset)             | None              | LlmConceptExtractor(Arc::clone main)   |
-/// | non-stub     | (unset)             | some(M)           | LlmConceptExtractor(new backend same type, model M) |
-async fn build_extractor(
-    main_backend: Arc<dyn InferenceBackend>,
-    main_backend_name: &str,
-    main_model: &str,
-    params: &BackendParams,
-    settings: ExtractorSettings,
-) -> Result<Arc<dyn ConceptExtractor>> {
-    // Warn on a flag combination the user probably didn't mean: passing
-    // --extractor-model alongside an explicit --extractor-backend stub
-    // looks like the model name will take effect, but stub ignores it.
-    if matches!(params.extractor_backend.as_deref(), Some("stub"))
-        && params.extractor_model.is_some()
-    {
-        tracing::warn!(
-            model = ?params.extractor_model,
-            "--extractor-model ignored: --extractor-backend is stub (deterministic, no model)"
-        );
-    }
-
-    match (main_backend_name, params.extractor_backend.as_deref()) {
-        (_, Some("stub")) => Ok(Arc::new(StubConceptExtractor::new())),
-        ("stub", None) => Ok(Arc::new(StubConceptExtractor::new())),
-        (_, Some(ext_backend_name)) => {
-            let model = params.extractor_model.clone().ok_or_else(|| {
-                PrimerError::Inference(
-                    "--extractor-model is required when --extractor-backend is set to a non-stub backend".into(),
-                )
-            })?;
-            let ext_backend = build_backend(ext_backend_name, model.clone(), params).await?;
-            Ok(Arc::new(LlmConceptExtractor::new(
-                ext_backend,
-                model,
-                settings,
-            )))
-        }
-        (_, None) => match params.extractor_model.as_deref() {
-            None => Ok(Arc::new(LlmConceptExtractor::new(
-                Arc::clone(&main_backend),
-                main_model.to_string(),
-                settings,
-            ))),
-            Some(override_model) => {
-                let ext_backend =
-                    build_backend(main_backend_name, override_model.to_string(), params).await?;
-                Ok(Arc::new(LlmConceptExtractor::new(
-                    ext_backend,
-                    override_model.to_string(),
-                    settings,
-                )))
-            }
-        },
-    }
-}
-
-/// Construct the comprehension classifier according to the same dispatch matrix
-/// as `build_extractor`:
-///
-/// | main backend | --comprehension-backend | --comprehension-model | outcome                                    |
-/// |--------------|-------------------------|-----------------------|--------------------------------------------|
-/// | stub         | (unset)                 | (any)                 | StubComprehensionClassifier                |
-/// | *            | "stub"                  | (any)                 | StubComprehensionClassifier                |
-/// | *            | some(X)                 | None                  | error (model required for non-stub)        |
-/// | *            | some(X)                 | some(M)               | LlmComprehensionClassifier(new backend X, model M) |
-/// | non-stub     | (unset)                 | None                  | LlmComprehensionClassifier(Arc::clone main) |
-/// | non-stub     | (unset)                 | some(M)               | LlmComprehensionClassifier(new backend same type, model M) |
-async fn build_comprehension(
-    main_backend: Arc<dyn InferenceBackend>,
-    main_backend_name: &str,
-    main_model: &str,
-    params: &BackendParams,
-    settings: primer_comprehension::ComprehensionSettings,
-) -> Result<Arc<dyn primer_comprehension::ComprehensionClassifier>> {
-    if matches!(params.comprehension_backend.as_deref(), Some("stub"))
-        && params.comprehension_model.is_some()
-    {
-        tracing::warn!(
-            model = ?params.comprehension_model,
-            "--comprehension-model ignored: --comprehension-backend is stub (deterministic, no model)"
-        );
-    }
-
-    match (main_backend_name, params.comprehension_backend.as_deref()) {
-        (_, Some("stub")) => Ok(Arc::new(
-            primer_comprehension::StubComprehensionClassifier::new(),
-        )),
-        ("stub", None) => Ok(Arc::new(
-            primer_comprehension::StubComprehensionClassifier::new(),
-        )),
-        (_, Some(comp_backend_name)) => {
-            let model = params.comprehension_model.clone().ok_or_else(|| {
-                PrimerError::Inference(
-                    "--comprehension-model is required when --comprehension-backend is set to a non-stub backend".into(),
-                )
-            })?;
-            let comp_backend = build_backend(comp_backend_name, model.clone(), params).await?;
-            Ok(Arc::new(
-                primer_comprehension::LlmComprehensionClassifier::new(
-                    comp_backend,
-                    model,
-                    settings,
-                ),
-            ))
-        }
-        (_, None) => match params.comprehension_model.as_deref() {
-            None => Ok(Arc::new(
-                primer_comprehension::LlmComprehensionClassifier::new(
-                    Arc::clone(&main_backend),
-                    main_model.to_string(),
-                    settings,
-                ),
-            )),
-            Some(override_model) => {
-                let comp_backend =
-                    build_backend(main_backend_name, override_model.to_string(), params).await?;
-                Ok(Arc::new(
-                    primer_comprehension::LlmComprehensionClassifier::new(
-                        comp_backend,
-                        override_model.to_string(),
-                        settings,
-                    ),
-                ))
-            }
-        },
-    }
-}
-
-/// Construct a fastembed-rs-backed `Embedder`. Returns `None` and emits
-/// a stderr error when the `embedding` cargo feature is not compiled
-/// in or when fastembed init fails. Failure is *not* fatal — the
-/// caller falls back to BM25-only retrieval.
-#[cfg(feature = "embedding")]
-fn build_fastembed_embedder(
-    model: Option<&str>,
-) -> Option<Arc<dyn primer_core::embedder::Embedder>> {
-    use primer_embedding::FastEmbedBackend;
-    let m = model.unwrap_or(primer_embedding::BGE_M3_MODEL_ID);
-    if m != primer_embedding::BGE_M3_MODEL_ID {
-        eprintln!(
-            "Note: --embedder-model {m} not yet supported by the CLI dispatch; using bge-m3."
-        );
-    }
-    eprintln!(
-        "Loading fastembed model {m}; first run downloads ~570 MB into ~/.cache/primer/models/."
-    );
-    match FastEmbedBackend::new() {
-        Ok(b) => Some(Arc::new(b) as _),
-        Err(e) => {
-            eprintln!("fastembed init failed ({e}); falling back to BM25-only retrieval.");
-            None
-        }
-    }
-}
-
-#[cfg(not(feature = "embedding"))]
-fn build_fastembed_embedder(
-    _model: Option<&str>,
-) -> Option<Arc<dyn primer_core::embedder::Embedder>> {
-    eprintln!(
-        "Error: --embedder-backend fastembed requires the `embedding` cargo feature. \
-         Build with `cargo run --features primer-cli/embedding -- ...` (or use --embedder-backend none)."
-    );
-    std::process::exit(1);
-}
-
-#[cfg(feature = "ollama-embedding")]
-async fn build_ollama_embedder(
-    url: Option<&str>,
-    model: Option<&str>,
-) -> Option<Arc<dyn primer_core::embedder::Embedder>> {
-    use primer_embedding::{DEFAULT_OLLAMA_MODEL, DEFAULT_OLLAMA_URL, OllamaEmbedder};
-    let url = url.unwrap_or(DEFAULT_OLLAMA_URL);
-    let model = model.unwrap_or(DEFAULT_OLLAMA_MODEL);
-    match OllamaEmbedder::with_endpoint(url, model).await {
-        Ok(b) => {
-            eprintln!("Embedder: ollama {model} at {url}");
-            Some(Arc::new(b) as _)
-        }
-        Err(e) => {
-            eprintln!("ollama embedder init failed ({e}); falling back to BM25-only retrieval.");
-            None
-        }
-    }
-}
-
-#[cfg(not(feature = "ollama-embedding"))]
-async fn build_ollama_embedder(
-    _url: Option<&str>,
-    _model: Option<&str>,
-) -> Option<Arc<dyn primer_core::embedder::Embedder>> {
-    eprintln!(
-        "Error: --embedder-backend ollama requires the `ollama-embedding` cargo feature. \
-         Build with `cargo run --features primer-cli/ollama-embedding -- ...`"
-    );
-    std::process::exit(1);
-}
-
-/// Reconcile a freshly-loaded persisted `LearnerModel` against the
-/// CLI flags for this launch.
-///
-/// Behaviour (kept minimal so the test surface matches the production
-/// branch exactly):
-/// - If `cli_name` differs from the persisted name, log a `tracing::warn!`
-///   AND a stderr `eprintln!` (so a parent who typos a name sees it
-///   without `RUST_LOG=warn`). The persisted name **always** wins —
-///   silently rewriting it would lock a child out of their own data.
-/// - Update `age` from the CLI (covers the birthday case).
-/// - Update `last_active` to now.
-///
-/// Returns the reconciled `LearnerModel`. The caller is responsible
-/// for the subsequent `save_learner` call (so this helper has no I/O).
-fn reconcile_persisted_learner(
-    mut existing: LearnerModel,
-    cli_name: &str,
-    cli_age: u8,
-) -> LearnerModel {
-    if existing.profile.name != cli_name {
-        eprintln!(
-            "Note: --name {:?} differs from the persisted learner name {:?}; \
-             keeping persisted (delete ~/.primer/<slug>.db to start fresh).",
-            cli_name, existing.profile.name
-        );
-        tracing::warn!(
-            "CLI --name {:?} differs from persisted learner name {:?}; using persisted",
-            cli_name,
-            existing.profile.name
-        );
-    }
-    existing.profile.age = cli_age;
-    existing.profile.last_active = Utc::now();
-    existing
-}
-
-/// Verify that a `--resume`'d session's stored locale matches the one
-/// requested at the CLI. Mismatches are a hard error: the session store
-/// has already opened with `cli_locale`, so any new concept inserts in
-/// the resumed session would be tagged with the wrong
-/// `concept_language_tag` — silent corruption of the longitudinal
-/// learner data, the kind of drift the project treats as a fail-fast
-/// condition for categorical state.
-///
-/// Returns `Ok(())` on match (or when the locales are equal) and
-/// `Err(message)` with a user-facing actionable string on mismatch.
-/// Pure for testability.
-fn verify_resume_locale_match(
-    cli_locale: Locale,
-    learner_locale: Locale,
-    resume_id: Uuid,
-) -> std::result::Result<(), String> {
-    if cli_locale == learner_locale {
-        return Ok(());
-    }
-    Err(format!(
-        "--resume {resume_id} was created in locale '{learner}', but \
-         --language '{cli}' was specified.\n  \
-         Drop --language to use the session's locale, or pass \
-         --language {learner}.",
-        learner = learner_locale.pack_id(),
-        cli = cli_locale.pack_id(),
-    ))
-}
 
 #[cfg(feature = "speech")]
 fn validate_speech_assets(
@@ -741,7 +264,7 @@ fn validate_speech_assets(
     voice_onnx: &Path,
     voice_config: &Path,
     voice_id: &str,
-) -> Result<()> {
+) -> primer_core::error::Result<()> {
     if !whisper_model.exists() {
         return Err(PrimerError::Speech(format!(
             "whisper model not found at {}.\n\
@@ -781,28 +304,6 @@ fn validate_speech_assets(
     Ok(())
 }
 
-fn create_learner_with_id(
-    id: Uuid,
-    name: &str,
-    age: u8,
-    locale: primer_core::i18n::Locale,
-) -> LearnerModel {
-    LearnerModel {
-        profile: LearnerProfile {
-            id,
-            name: name.to_string(),
-            age,
-            languages: vec![locale.pack_id().to_string()],
-            locale,
-            created_at: Utc::now(),
-            last_active: Utc::now(),
-        },
-        concepts: vec![],
-        preferences: LearningPreferences::default(),
-        current_engagement: EngagementState::Engaged,
-        recent_assessments: vec![],
-    }
-}
 
 /// Probe common system locations for an `espeak-ng-data` directory and
 /// set `PIPER_ESPEAKNG_DATA_DIRECTORY` to the parent of the first complete
@@ -1412,354 +913,6 @@ async fn async_main() -> anyhow::Result<()> {
     Ok(())
 }
 
-#[cfg(test)]
-mod classifier_construction_tests {
-    use super::*;
-
-    /// Build a minimal `BackendParams` for testing.
-    fn params_with(
-        classifier_backend: Option<&str>,
-        classifier_model: Option<&str>,
-    ) -> BackendParams {
-        BackendParams {
-            api_key: None,
-            ollama_url: "http://localhost:11434".into(),
-            classifier_backend: classifier_backend.map(String::from),
-            classifier_model: classifier_model.map(String::from),
-            extractor_backend: None,
-            extractor_model: None,
-            comprehension_backend: None,
-            comprehension_model: None,
-        }
-    }
-
-    /// With main=stub and no classifier flags, we get a stub classifier.
-    #[tokio::test]
-    async fn stub_main_no_flags_gives_stub_classifier() {
-        let params = params_with(None, None);
-        let main = build_backend("stub", "main".into(), &params).await.unwrap();
-        let c = build_classifier(main, "stub", "main", &params, ClassifierSettings::default())
-            .await
-            .unwrap();
-        assert_eq!(c.identifier(), "stub");
-    }
-
-    /// `--classifier-backend stub` wins regardless of the main backend.
-    #[tokio::test]
-    async fn explicit_stub_override_wins_over_any_main() {
-        // Main is stub (would default to stub anyway), but the explicit flag
-        // should also work when the main is declared as a different type.
-        // We use stub here to avoid needing a real cloud backend in tests.
-        let params_main = params_with(None, None);
-        let main = build_backend("stub", "main-model".into(), &params_main)
-            .await
-            .unwrap();
-        let params = params_with(Some("stub"), None);
-        let c = build_classifier(
-            main,
-            "cloud",
-            "main-model",
-            &params,
-            ClassifierSettings::default(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(c.identifier(), "stub");
-    }
-
-    /// No classifier flags on a non-stub main → LlmEngagementClassifier wrapping
-    /// Arc::clone of main (identifier is "llm:<main-model>").
-    #[tokio::test]
-    async fn non_stub_main_no_flags_gives_llm_with_main_model() {
-        // We simulate a non-stub main by using "stub" type but treating it as
-        // cloud in the dispatch (the type doesn't matter for the Arc::clone path;
-        // what matters is that main_backend_name is NOT "stub").
-        // Since we can't construct a real cloud backend without an API key,
-        // we build a stub backend but pass "cloud" as the name to exercise
-        // the dispatch arm that reuses main via Arc::clone.
-        let params = params_with(None, None);
-        let main = build_backend("stub", "claude-sonnet-4-6".into(), &params)
-            .await
-            .unwrap();
-        let c = build_classifier(
-            main,
-            "cloud", // use "cloud" as name to exercise the non-stub arm
-            "claude-sonnet-4-6",
-            &params,
-            ClassifierSettings::default(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(c.identifier(), "llm:claude-sonnet-4-6");
-    }
-
-    /// `--classifier-backend ollama` without `--classifier-model` is an error.
-    #[tokio::test]
-    async fn classifier_backend_without_model_is_error() {
-        let params_main = params_with(None, None);
-        let main = build_backend("stub", "main".into(), &params_main)
-            .await
-            .unwrap();
-        let params = params_with(Some("ollama"), None /* no model */);
-        let result =
-            build_classifier(main, "stub", "main", &params, ClassifierSettings::default()).await;
-        assert!(
-            result.is_err(),
-            "should error when --classifier-backend is non-stub and --classifier-model is missing"
-        );
-    }
-
-    /// `--classifier-model` override on a stub main still yields stub classifier
-    /// (the stub arm fires before we reach the model-override arm).
-    #[tokio::test]
-    async fn classifier_model_override_on_stub_main_still_yields_stub() {
-        let params = params_with(None, Some("override-model"));
-        let main = build_backend("stub", "main".into(), &params).await.unwrap();
-        let c = build_classifier(main, "stub", "main", &params, ClassifierSettings::default())
-            .await
-            .unwrap();
-        // Main is stub → classifier defaults to stub regardless of model override.
-        assert_eq!(c.identifier(), "stub");
-    }
-
-    /// `--classifier-model` override on a NON-stub main constructs a fresh
-    /// backend of the same type with the override model (case 5 in the dispatch
-    /// matrix, lines 284-292 in build_classifier).
-    ///
-    /// The `main_backend` Arc is a stub (we cannot unit-test real cloud/ollama
-    /// without live infrastructure), but `main_backend_name` is "ollama" so the
-    /// dispatch falls through to the model-override branch and calls
-    /// `build_backend("ollama", "override-model", params)`.  `OllamaBackend::new`
-    /// is purely constructive (no I/O), so this succeeds in unit tests.
-    #[tokio::test]
-    async fn classifier_model_override_on_non_stub_main_constructs_fresh_backend() {
-        let params = params_with(None, Some("override-model"));
-        // The Arc itself is a stub, but the *name* string "ollama" drives dispatch.
-        let main = build_backend("stub", "main".into(), &params).await.unwrap();
-        let c = build_classifier(
-            main,
-            "ollama", // non-stub name → falls through to model-override arm
-            "main-model",
-            &params,
-            ClassifierSettings::default(),
-        )
-        .await
-        .unwrap();
-        // A fresh OllamaBackend is constructed with "override-model"; the
-        // classifier identifier must reflect the override, not the main model.
-        assert_eq!(c.identifier(), "llm:override-model");
-    }
-
-    /// An unknown `--classifier-backend` value must return an error (the
-    /// explicit-backend arm calls `build_backend`, which errors on unknown names).
-    #[tokio::test]
-    async fn unknown_classifier_backend_returns_error() {
-        let params = params_with(Some("nonexistent"), Some("any-model"));
-        let main = build_backend("stub", "main".into(), &params).await.unwrap();
-        let result =
-            build_classifier(main, "stub", "main", &params, ClassifierSettings::default()).await;
-        assert!(
-            result.is_err(),
-            "should error when --classifier-backend names an unknown backend"
-        );
-    }
-}
-
-#[cfg(test)]
-mod extractor_construction_tests {
-    use super::*;
-
-    fn params(extractor_backend: Option<&str>, extractor_model: Option<&str>) -> BackendParams {
-        BackendParams {
-            api_key: Some("k".into()),
-            ollama_url: "http://localhost:11434".into(),
-            classifier_backend: None,
-            classifier_model: None,
-            extractor_backend: extractor_backend.map(String::from),
-            extractor_model: extractor_model.map(String::from),
-            comprehension_backend: None,
-            comprehension_model: None,
-        }
-    }
-
-    fn stub_main_backend() -> Arc<dyn InferenceBackend> {
-        Arc::new(primer_inference::stub::StubBackend)
-    }
-
-    #[tokio::test]
-    async fn stub_main_no_flags_gives_stub_extractor() {
-        let e = build_extractor(
-            stub_main_backend(),
-            "stub",
-            "stub",
-            &params(None, None),
-            ExtractorSettings::default(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(e.identifier(), "stub");
-    }
-
-    #[tokio::test]
-    async fn explicit_stub_override_wins() {
-        let e = build_extractor(
-            stub_main_backend(),
-            "ollama",
-            "llama3.2",
-            &params(Some("stub"), None),
-            ExtractorSettings::default(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(e.identifier(), "stub");
-    }
-
-    #[tokio::test]
-    async fn nonstub_backend_without_model_errors() {
-        let r = build_extractor(
-            stub_main_backend(),
-            "ollama",
-            "llama3.2",
-            &params(Some("ollama"), None),
-            ExtractorSettings::default(),
-        )
-        .await;
-        assert!(r.is_err());
-    }
-
-    // Note: `stub_main_backend()` always returns a StubBackend (whose
-    // `name()` is "stub"), even when `main_backend_name` says otherwise.
-    // The identifier therefore reads as `llm:stub:<model>` in these
-    // tests — what they actually verify is that the *model* name flows
-    // through correctly. In production the main_backend_name and the
-    // backend's own `name()` agree by construction.
-    #[tokio::test]
-    async fn no_flags_reuses_main_model_id() {
-        let e = build_extractor(
-            stub_main_backend(),
-            "ollama",
-            "llama3.2",
-            &params(None, None),
-            ExtractorSettings::default(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(e.identifier(), "llm:stub:llama3.2");
-    }
-
-    #[tokio::test]
-    async fn model_only_override_uses_main_backend_type() {
-        // With `--extractor-model haiku` and main backend `ollama`,
-        // `build_backend("ollama", "haiku", ...)` constructs a real
-        // OllamaBackend, so the identifier carries the real backend
-        // name (unlike the `_reuses_main_model_id` test above which
-        // Arc::clones the stub harness).
-        let e = build_extractor(
-            stub_main_backend(),
-            "ollama",
-            "llama3.2",
-            &params(None, Some("haiku")),
-            ExtractorSettings::default(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(e.identifier(), "llm:ollama:haiku");
-    }
-}
-
-#[cfg(test)]
-mod comprehension_construction_tests {
-    use super::*;
-
-    fn params(
-        comprehension_backend: Option<&str>,
-        comprehension_model: Option<&str>,
-    ) -> BackendParams {
-        BackendParams {
-            api_key: Some("k".into()),
-            ollama_url: "http://localhost:11434".into(),
-            classifier_backend: None,
-            classifier_model: None,
-            extractor_backend: None,
-            extractor_model: None,
-            comprehension_backend: comprehension_backend.map(String::from),
-            comprehension_model: comprehension_model.map(String::from),
-        }
-    }
-
-    fn stub_main_backend() -> Arc<dyn InferenceBackend> {
-        Arc::new(primer_inference::stub::StubBackend)
-    }
-
-    #[tokio::test]
-    async fn stub_main_no_flags_gives_stub_comprehension() {
-        let c = build_comprehension(
-            stub_main_backend(),
-            "stub",
-            "stub",
-            &params(None, None),
-            primer_comprehension::ComprehensionSettings::default(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(c.identifier(), "stub");
-    }
-
-    #[tokio::test]
-    async fn explicit_stub_override_wins() {
-        let c = build_comprehension(
-            stub_main_backend(),
-            "ollama",
-            "llama3.2",
-            &params(Some("stub"), None),
-            primer_comprehension::ComprehensionSettings::default(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(c.identifier(), "stub");
-    }
-
-    #[tokio::test]
-    async fn nonstub_backend_without_model_errors() {
-        let r = build_comprehension(
-            stub_main_backend(),
-            "ollama",
-            "llama3.2",
-            &params(Some("ollama"), None),
-            primer_comprehension::ComprehensionSettings::default(),
-        )
-        .await;
-        assert!(r.is_err());
-    }
-
-    #[tokio::test]
-    async fn no_flags_reuses_main_model_id() {
-        let c = build_comprehension(
-            stub_main_backend(),
-            "ollama",
-            "llama3.2",
-            &params(None, None),
-            primer_comprehension::ComprehensionSettings::default(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(c.identifier(), "llm:stub:llama3.2");
-    }
-
-    #[tokio::test]
-    async fn model_only_override_uses_main_backend_type() {
-        let c = build_comprehension(
-            stub_main_backend(),
-            "ollama",
-            "llama3.2",
-            &params(None, Some("haiku")),
-            primer_comprehension::ComprehensionSettings::default(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(c.identifier(), "llm:ollama:haiku");
-    }
-}
 
 #[cfg(test)]
 mod tests {
@@ -1790,127 +943,6 @@ mod tests {
         assert!(parse_mic_silence_ms("-100").is_err());
     }
 
-    #[test]
-    fn slug_lowercases_and_keeps_alphanumerics() {
-        assert_eq!(slug("Explorer"), "explorer");
-        assert_eq!(slug("Binti7"), "binti7");
-    }
-
-    #[test]
-    fn slug_replaces_special_chars_with_dash() {
-        assert_eq!(slug("Anna Maria"), "anna-maria");
-        assert_eq!(slug("Lee/Davis"), "lee-davis");
-    }
-
-    #[test]
-    fn slug_keeps_unicode_letters_lowercased() {
-        // The previous ASCII-only rule collapsed `José`, `Łukasz`, `Соня`
-        // and `美咲` into ambiguous or empty stems. Children's names are
-        // a load-bearing input here — accept anything Unicode considers
-        // alphanumeric, lowercased where that exists.
-        assert_eq!(slug("José"), "josé");
-        assert_eq!(slug("Łukasz"), "łukasz");
-        assert_eq!(slug("Соня"), "соня");
-        // No case-folding for CJK; the chars round-trip as-is.
-        assert_eq!(slug("美咲"), "美咲");
-    }
-
-    #[test]
-    fn slug_normalizes_nfc_so_decomposed_equals_precomposed() {
-        // Same visible name, two Unicode encodings: precomposed `é`
-        // (U+00E9) vs decomposed `e` + combining acute (U+0301). Without
-        // NFC normalization these slug to different filenames, so two
-        // copies of the same child get two session DBs.
-        let nfc = "Jos\u{00E9}"; // José (NFC)
-        let nfd = "Jose\u{0301}"; // José (NFD)
-        assert_eq!(slug(nfc), slug(nfd));
-    }
-
-    #[test]
-    fn slug_collapses_runs_of_separators() {
-        assert_eq!(slug("a   b"), "a-b");
-        assert_eq!(slug("a---b"), "a-b");
-    }
-
-    #[test]
-    fn slug_strips_leading_and_trailing_separators() {
-        assert_eq!(slug("  hello  "), "hello");
-        assert_eq!(slug("___world___"), "world");
-    }
-
-    #[test]
-    fn slug_empty_input_falls_back_to_default() {
-        assert_eq!(slug(""), "default");
-        assert_eq!(slug("!!!"), "default");
-    }
-
-    #[test]
-    fn verify_resume_locale_match_ok_when_locales_equal() {
-        let id = Uuid::new_v4();
-        assert!(verify_resume_locale_match(Locale::English, Locale::English, id).is_ok());
-        assert!(verify_resume_locale_match(Locale::German, Locale::German, id).is_ok());
-    }
-
-    #[test]
-    fn verify_resume_locale_match_errors_on_mismatch_with_actionable_message() {
-        let id = Uuid::new_v4();
-        let err = verify_resume_locale_match(Locale::English, Locale::German, id).unwrap_err();
-        assert!(
-            err.contains("'de'"),
-            "must name the learner's locale: {err}"
-        );
-        assert!(err.contains("'en'"), "must name the cli locale: {err}");
-        assert!(
-            err.contains("--language de"),
-            "must show the corrective flag value: {err}"
-        );
-        assert!(
-            err.contains(&id.to_string()),
-            "must include the resume id so the user knows which session: {err}"
-        );
-    }
-
-    #[test]
-    fn verify_resume_locale_match_symmetric_de_to_en() {
-        let id = Uuid::new_v4();
-        let err = verify_resume_locale_match(Locale::German, Locale::English, id).unwrap_err();
-        assert!(
-            err.contains("'en'"),
-            "must name the learner's locale: {err}"
-        );
-        assert!(err.contains("--language en"), "corrective flag: {err}");
-    }
-
-    #[test]
-    fn resolve_session_db_path_passes_explicit_through() {
-        // The home arg is unused when an explicit path is given.
-        let home = Path::new("/this/should/be/ignored");
-        let p = resolve_session_db_path(
-            Some(PathBuf::from("/tmp/explicit.db")),
-            home,
-            "Anyone",
-            false,
-        );
-        assert_eq!(p, PathBuf::from("/tmp/explicit.db"));
-    }
-
-    #[test]
-    fn resolve_session_db_path_default_uses_home_and_slug() {
-        let home = Path::new("/synthetic/home");
-        let p = resolve_session_db_path(None, home, "Binti", false);
-        assert_eq!(p, PathBuf::from("/synthetic/home/.primer/binti.db"));
-    }
-
-    #[test]
-    fn resolve_session_db_path_no_persist_returns_in_memory() {
-        // `--no-persist` short-circuits everything: no slug, no home
-        // join, no explicit path. The session is throwaway.
-        let home = Path::new("/some/home");
-        assert_eq!(
-            resolve_session_db_path(None, home, "Anyone", true),
-            PathBuf::from(IN_MEMORY)
-        );
-    }
 
     #[test]
     fn no_persist_conflicts_with_resume_at_parse_time() {
@@ -1932,134 +964,6 @@ mod tests {
         // contradiction; clap should reject it up front.
         let result = Cli::try_parse_from(["primer", "--no-persist", "--session-db", "/tmp/x.db"]);
         assert!(result.is_err(), "expected clap to reject the combination");
-    }
-
-    #[test]
-    fn first_run_banner_shows_only_for_default_path_first_run() {
-        // Default path + brand-new file → show banner (the user just
-        // started persisting without explicitly opting in).
-        assert!(should_show_first_run_banner(false, false, false));
-        // Default path but file already existed → silent (not first run).
-        assert!(!should_show_first_run_banner(false, false, true));
-        // Explicit path → silent (the user knows where their data is).
-        assert!(!should_show_first_run_banner(true, false, false));
-        // No-persist → silent (no file is being created at all).
-        assert!(!should_show_first_run_banner(false, true, false));
-    }
-
-    #[test]
-    fn resolve_session_db_path_default_handles_unicode_name() {
-        // Confirms the slug + path composition round-trip a non-ASCII
-        // name without env mutation. The same name in NFC vs NFD must
-        // produce the same path so we don't end up with two DB files.
-        let home = Path::new("/h");
-        assert_eq!(
-            resolve_session_db_path(None, home, "José", false),
-            PathBuf::from("/h/.primer/josé.db")
-        );
-        let nfd = "Jose\u{0301}";
-        assert_eq!(
-            resolve_session_db_path(None, home, nfd, false),
-            PathBuf::from("/h/.primer/josé.db")
-        );
-    }
-
-    #[tokio::test]
-    async fn cli_birthday_case_updates_age_and_keeps_uuid() {
-        // Save a learner with age=8, simulate startup with --age=9 by
-        // calling the SAME helper main() uses, then verify the persisted
-        // row has age=9 with the same UUID and created_at preserved.
-        use chrono::Utc;
-        use primer_core::storage::LearnerStore;
-        use primer_storage::SqliteSessionStore;
-        use std::sync::Arc;
-
-        let store = Arc::new(
-            SqliteSessionStore::open_for_locale(
-                std::path::Path::new(":memory:"),
-                primer_core::i18n::Locale::default(),
-            )
-            .unwrap(),
-        );
-        let original_id = Uuid::new_v4();
-        let original_created = Utc::now() - chrono::Duration::days(365);
-        let mut original =
-            create_learner_with_id(original_id, "Binti", 8, primer_core::i18n::Locale::English);
-        original.profile.created_at = original_created;
-        store.save_learner(&original).await.unwrap();
-
-        // Reload + reconcile via the production helper.
-        let existing = store.load_learner().await.unwrap().expect("learner row");
-        let reconciled = reconcile_persisted_learner(existing, "Binti", 9);
-        store.save_learner(&reconciled).await.unwrap();
-
-        assert_eq!(
-            reconciled.profile.id, original_id,
-            "UUID stable across launches"
-        );
-        assert_eq!(reconciled.profile.age, 9, "age updated to CLI value");
-        assert_eq!(
-            reconciled.profile.created_at.timestamp(),
-            original_created.timestamp(),
-            "created_at preserved",
-        );
-    }
-
-    #[tokio::test]
-    async fn cli_name_mismatch_keeps_persisted_name() {
-        // Save with name="Binti", call reconcile_persisted_learner with
-        // --name="Other" — the SAME helper main() uses — and verify the
-        // persisted name stays "Binti". The tracing::warn! / eprintln!
-        // emission is intentionally NOT asserted (subscriber capture
-        // would over-couple the test); the data invariant is what
-        // matters here, and exercising the production helper proves we
-        // are testing the actual production branch rather than a stub.
-        use primer_core::storage::LearnerStore;
-        use primer_storage::SqliteSessionStore;
-        use std::sync::Arc;
-
-        let store = Arc::new(
-            SqliteSessionStore::open_for_locale(
-                std::path::Path::new(":memory:"),
-                primer_core::i18n::Locale::default(),
-            )
-            .unwrap(),
-        );
-        let original = create_learner_with_id(
-            Uuid::new_v4(),
-            "Binti",
-            8,
-            primer_core::i18n::Locale::English,
-        );
-        store.save_learner(&original).await.unwrap();
-
-        let existing = store.load_learner().await.unwrap().expect("learner row");
-        let reconciled = reconcile_persisted_learner(existing, "Other", 8);
-        store.save_learner(&reconciled).await.unwrap();
-
-        assert_eq!(
-            reconciled.profile.name, "Binti",
-            "persisted name wins over CLI"
-        );
-
-        // Round-trip through the store too — proves the saved row also
-        // keeps the persisted name (i.e. the helper didn't mutate name
-        // before save_learner committed it).
-        let round_trip = store.load_learner().await.unwrap().expect("learner row");
-        assert_eq!(round_trip.profile.name, "Binti");
-    }
-
-    #[test]
-    fn reconcile_persisted_learner_preserves_name_and_id_on_match() {
-        // The non-mismatch path: same name should be a pure age/last_active
-        // refresh with no warn (covered by absence of stderr in this test).
-        let original_id = Uuid::new_v4();
-        let original =
-            create_learner_with_id(original_id, "Binti", 8, primer_core::i18n::Locale::English);
-        let result = reconcile_persisted_learner(original, "Binti", 9);
-        assert_eq!(result.profile.name, "Binti");
-        assert_eq!(result.profile.id, original_id);
-        assert_eq!(result.profile.age, 9);
     }
 
     // ─── --vocab-max-per-prompt parse tests ─────────────────────────────
